% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_palette.R
\name{print.palette}
\alias{print.palette}
\title{Function for printing palette}
\usage{
\method{print}{palette}(x, ...)
}
\arguments{
\item{x}{the object to be printed.}

\item{...}{further arguments to be passed to or from other methods.
They are ignored in this function.}
}
\value{
A plot of the specified colour palette.
}
\description{
Function for printing palette
}
\examples{
print(RSScols("signif_qual"))
}
