% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSPalettes.R
\docType{data}
\name{RSSPalettes}
\alias{RSSPalettes}
\title{Complete list of palettes.}
\format{
An object of class \code{list} of length 3.
}
\usage{
RSSPalettes
}
\description{
Use \code{names(RSSPalettes)} to return all palette names.
}
\keyword{datasets}
