% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_rss_c.R
\name{scale_fill_rss_c}
\alias{scale_fill_rss_c}
\title{Plotting with RSS palettes for fill with ggplot2}
\usage{
scale_fill_rss_c(palette, direction = 1, ...)
}
\arguments{
\item{palette}{name of palette. Run \code{names(RSSPalettes)} to view options.}

\item{direction}{Sets order of colors. Default palette_choice is 1. If direction is -1,
palette_choice color order is reversed}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{scale_fill_gradientn}}}
}
\value{
A ggproto object defining a continuous colour scale for use with ggplot2.
}
\description{
Plotting with RSS palettes for fill with ggplot2
}
\examples{
library(ggplot2)
ggplot(data = mtcars, aes(x = cyl, y = disp, fill = cyl)) +
  geom_col() +
  scale_fill_rss_c(palette = "signif_seq")
}
