% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rss_palette.R
\name{set_rss_palette}
\alias{set_rss_palette}
\title{Set RSS base R plotting palette}
\usage{
set_rss_palette(palette, ...)
}
\arguments{
\item{palette}{Name of palette. See \code{names(RSSthemes::RSSPalettes)}.}
}
\description{
Set RSS base R plotting palette
}
