% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSPalettes.R
\docType{data}
\name{signif_blue}
\alias{signif_blue}
\title{Significance blue hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
signif_blue
}
\description{
Significance blue hex colour
}
\keyword{datasets}
