% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSPalettes.R
\docType{data}
\name{signif_green}
\alias{signif_green}
\title{Significance green hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
signif_green
}
\description{
Significance green hex colour
}
\keyword{datasets}
