% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSPalettes.R
\docType{data}
\name{signif_orange}
\alias{signif_orange}
\title{Significance orange hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
signif_orange
}
\description{
Significance orange hex colour
}
\keyword{datasets}
