% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSPalettes.R
\docType{data}
\name{signif_red}
\alias{signif_red}
\title{Significance red hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
signif_red
}
\description{
Significance red hex colour
}
\keyword{datasets}
