% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSPalettes.R
\docType{data}
\name{signif_yellow}
\alias{signif_yellow}
\title{Significance yellow hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
signif_yellow
}
\description{
Significance yellow hex colour
}
\keyword{datasets}
