% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_significance.R
\name{theme_significance}
\alias{theme_significance}
\title{Significance theme}
\usage{
theme_significance(
  main_font = "sans",
  title_font = "sans",
  hjust = 0,
  text_col = "black",
  bg_col = "white"
)
}
\arguments{
\item{main_font}{Font used for all text elements except the title. Default "sans".}

\item{title_font}{Font used for the title. Default "sans".}

\item{hjust}{Horizontal alignment of title, subtitle, and caption. Default 0.}

\item{text_col}{Text colour. Default "black".}

\item{bg_col}{Background colour. Default "white".}
}
\value{
A ggplot2 theme
}
\description{
Custom ggplot theme
}
