% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_all_palettes.R
\name{view_all_palettes}
\alias{view_all_palettes}
\title{Prints all available colour palettes}
\usage{
view_all_palettes(colourblind_friendly = FALSE)
}
\arguments{
\item{colourblind_friendly}{Boolean whether or not to return only colourblind friendly
palettes. Default \code{FALSE}.}
}
\value{
A plot of all colour palettes available in the package.
}
\description{
Prints all available colour palettes
}
\examples{
view_all_palettes()
view_all_palettes(colourblind_friendly = TRUE)
}
