% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rss_palette.R
\name{set_rss_palette}
\alias{set_rss_palette}
\title{Set RSS base R plotting palette}
\usage{
set_rss_palette(palette)
}
\arguments{
\item{palette}{Name of palette. See \code{names(RSSthemes::RSSPalettes)}.}
}
\value{
Returns a character vector giving the colors from the palette which
was in effect. This is invisible unless the argument is omitted.
}
\description{
Set RSS base R plotting palette
}
\examples{
set_rss_palette("signif_qual")
}
