% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_significance.R
\name{theme_significance}
\alias{theme_significance}
\title{Significance theme}
\usage{
theme_significance(
  family = "Source Sans Pro",
  base_size = 13,
  hjust = 0,
  text_col = "black",
  bg_col = "white"
)
}
\arguments{
\item{family}{Font used for all text elements. Default "Source Sans Pro".}

\item{base_size}{Base font size for text elements. Default 12.}

\item{hjust}{Horizontal alignment of title, subtitle, and caption. Default 0.}

\item{text_col}{Text colour. Default "black".}

\item{bg_col}{Background colour. Default "white".}
}
\value{
A ggplot2 theme
}
\description{
Custom ggplot theme
}
\examples{
library(ggplot2)
ggplot(data = mtcars, aes(x = mpg, y = disp, color = wt)) +
  geom_point() +
  theme_significance()
}
